Move fetchTaxonAndSave back to original file
